requirements.txt
Flask==0.12.1
Flask-Cors==3.0.2
numpy==1.16.2
numpydoc==0.6.0
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
opencv-python==4.1.0.25
pyzbar==0.1.8
python-docx==0.8.6
regex==2018.08.17
oauth2client==4.1.2
google-api-python-client==1.7.4
httplib2==0.11.3
pandas==0.24.2
xlrd==1.1.0
sqlalchemy==1.3.1
pathlib2==2.3.3
wtforms==2.2.1
flask-sqlalchemy==2.4.0
docxtpl==0.6.2
python-barcode==0.10.0
imutils==0.5.2
pdf2image==1.6.0
nltk==3.4
gunicorn


